1 add entity classes (#9)
* update database scripts

* add new entities to db

* update all tests so they will use unique db

* add door db unit tests

* add role entity unit tests

* add migrations

* add role table unit tests

* add migration designer

* fix variable names

* add permission db unit tests

* rename event entity to incident entity

* add incident table unit tests

* update item groups in project file

* delete unnecessary authentication code

* exclude migration class from code coverage

* update event table name

* fix required db fields

* fix incident table unit tests

* exclude migrations from code coverage

* fix apply migration shell file

* fix apply migration shell file

* improve coverage for permission entity

* improve code coverage of user entity

+++ Libraries/KeyManager.Domain/Interfaces/IGenericRepository.cs
@@ -8,7 +8,7 @@ public interface IGenericRepository<T> where T : BaseEntity
 
     Task<List<T>> FindAsync(Expression<Func<T, bool>> expression);
 
-    Task<T> GetAsync(int id);
+    Task<T> GetAsync(long id);
 
     Task<T> GetAsync(Expression<Func<T, bool>> expression);
